+++ packages/events/src/service/indexing/utils.ts
@@ -476,6 +476,7 @@ export function resolveRecordActionScopeToIds(
       placeOfEvent: getLocationIdsFromScopeOptions(options?.placeOfEvent, user),
       createdBy:
         options?.createdBy === UserFilter.enum.user ? user.id : undefined,
+      createdIn: getLocationIdsFromScopeOptions(options?.createdIn, user),
       declaredIn: getLocationIdsFromScopeOptions(options?.declaredIn, user),
       declaredBy:
         options?.declaredBy === UserFilter.enum.user ? user.id : undefined,